Fix default Home preference FC on orientation change
When an orientation change is triggered, onRestoreInstanceState is called to make sure that
details about the current View and the List are saved and restored correctly (so the List
can become populated after the orientation change).

Current code was fine for Assist, Browser, Phone, SMS, Emergency default preferences because all
of them were populating the list with packageNames, and setPackageNames was called correctly
at onRestoreInstanceState.

But Home preference was setting componentNames, not packageNames, thus the state was never
restored correctly (because the passed componentNames were caught by the NameNotFoundException#L162)
and it was ending up with an IndexOutOfBoundsException on every orientation change.

Logcat: https://hastebin.com/avadosomah.pl

Steps to reproduce:
1. Open home preference dialog
2. Rotate the device
3. The list is empty
4. Tap anywhere to trigger the FC

This commit introduces a new method that checks the contect of the passed entryValues, decides whether
they are componentNames or packageNames and passes them accordingly.

Change-Id: I6ae64b641ff7cb17c9fc467b54ee0ed966942449

Signed-off-by: xyyx <xyyx@mail.ru>